restore consensus_test.go
