+++ Makefile
@@ -9,5 +9,5 @@ install:
 	go install github.com/avito-tech/go-mutesting/...
 
 mut:
-	go-mutesting --blacklist=".github/mut_blacklist" --config=".github/mut_config.yml" ./...
+	MUTATION_TEST=on go-mutesting --blacklist=".github/mut_blacklist" --config=".github/mut_config.yml" ./...
 	@echo MSI: `jq '.stats.msi' report.json`
